improve lottery drawing; improve player skill usage validation; partially fix, DoT effects and moaico; fix cloth equipment updating

+++ game-server/src/main/java/com/jftse/emulator/server/core/task/DespawnCrystalTask.java
@@ -52,7 +52,7 @@ public class DespawnCrystalTask extends AbstractTask {
                 !isBattleGame && (((MatchplayGuardianGame) game).getSkillCrystals().stream().anyMatch(x -> x.getId() == skillCrystal.getId()));
 
         if (isCrystalStillAvailable) {
-            S2CMatchplayLetCrystalDisappear letCrystalDisappearPacket = new S2CMatchplayLetCrystalDisappear(skillCrystal.getId());
+            S2CMatchplayLetCrystalDisappear letCrystalDisappearPacket = new S2CMatchplayLetCrystalDisappear((short) skillCrystal.getId());
             GameManager.getInstance().sendPacketToAllClientsInSameGameSession(letCrystalDisappearPacket, connection);
 
             ConcurrentLinkedDeque<SkillCrystal> skillCrystals = isBattleGame ? ((MatchplayBattleGame) game).getSkillCrystals() : ((MatchplayGuardianGame) game).getSkillCrystals();
